Split sort algorithms to multiple files
