+++ src/sort.c
@@ -1,9 +1,9 @@
-#include "sort.h"
+#include "sort-impl.h"
 #include <float.h>
 #include <math.h>
 
-static double get_scalar(struct array *a, uint i, uint offset, uint usize,
-                         gs_dom type) {
+double get_scalar(struct array *a, uint i, uint offset, uint usize,
+                  gs_dom type) {
   char *v = (char *)a->ptr + i * usize + offset;
 
   double data;
@@ -24,8 +24,8 @@ static double get_scalar(struct array *a, uint i, uint offset, uint usize,
   return data;
 }
 
-static void get_extrema(void *extrema_, struct sort *data, uint field,
-                        const struct comm *c) {
+void get_extrema(void *extrema_, struct sort *data, uint field,
+                 const struct comm *c) {
   struct array *a = data->a;
   uint usize = data->unit_size;
   uint offset = data->offset[field];
@@ -46,9 +46,10 @@ static void get_extrema(void *extrema_, struct sort *data, uint field,
   extrema[0] *= -1;
 }
 
-static int set_dest(uint *proc, uint size, sint np, slong start, slong nelem) {
+void set_proc_from_idx(uint *proc, uint size, sint np, slong start,
+                       slong nelem) {
   if (nelem == 0)
-    return 1;
+    return;
 
   uint nelt = nelem / np, nrem = nelem - np * nelt;
   if (nrem == 0) {
@@ -65,45 +66,6 @@ static int set_dest(uint *proc, uint size, sint np, slong start, slong nelem) {
         proc[i] = s + (start + i - t) / (nelt + 1);
     }
   }
-
-  return 0;
-}
-
-//-----------------------------------------------------------------------------
-// Parallel Bin-Sort
-//
-static uint *set_bin(struct sort *s, uint field, const struct comm *c) {
-  struct array *a = s->a;
-  gs_dom t = s->t[field];
-  uint offset = s->offset[field];
-
-  uint size = a->n;
-  uint *proc = tcalloc(uint, size + 1);
-
-  double extrema[2];
-  get_extrema((void *)extrema, s, field, c);
-  double range = extrema[1] - extrema[0];
-
-  if (size == 0)
-    return 0;
-
-  sint np = c->np;
-  uint id = 0;
-  uint index = 0;
-  do {
-    double end = extrema[0] + (range / np) * (id + 1);
-    while (index < size) {
-      double val = get_scalar(a, index, offset, s->unit_size, t);
-      if (val <= end)
-        proc[index] = id, index++;
-      else
-        break;
-    }
-    id++;
-  } while (id < np && index < size);
-  for (; index < size; index++)
-    proc[index] = np - 1;
-  return 0;
 }
 
 static int sort_field(struct array *arr, size_t usize, gs_dom t, uint off,
@@ -130,7 +92,7 @@ static int sort_field(struct array *arr, size_t usize, gs_dom t, uint off,
   return 0;
 }
 
-int sort_local(struct sort *s) {
+void sort_local(struct sort *s) {
   struct array *a = s->a;
   buffer *buf = s->buf;
   size_t usize = s->unit_size;
@@ -140,228 +102,6 @@ int sort_local(struct sort *s) {
   while (i >= 0)
     sort_field(a, usize, s->t[i], s->offset[i], buf, 1), i--;
   sarray_permute_buf_(s->align, usize, a->ptr, a->n, buf);
-
-  return 0;
-}
-
-static void parallel_bin_sort(struct sort *s, const struct comm *c) {
-  // Locally sort the array first.
-  sort_local(s);
-
-  // Calculate the global array size. If it is zero, nothing to do, just return.
-  struct array *arr = s->a;
-  slong ng = arr->n, wrk[2];
-  comm_allreduce(c, gs_long, gs_add, &ng, 1, wrk);
-  if (ng == 0)
-    return;
-
-  // Set destination bin based on the field value.
-  uint *proc = set_bin(s, 0, c);
-
-  // Initialize the crystal router.
-  struct crystal cr;
-  crystal_init(&cr, c);
-
-  // Transfer the array elements to destination processor. To avoid message
-  // sizes larger than INT_MAX, we calculate total message size and then figure
-  // out how many transfers we need. Then we transfer array using that many
-  // transfers.
-  size_t usize = s->unit_size;
-  uint nt = 2 * ((ng * usize + INT_MAX - 1) / INT_MAX);
-  uint tsize = (arr->n + nt - 1) / nt;
-
-  struct array brr, crr;
-  array_init_(&brr, tsize + 1, usize, __FILE__, __LINE__);
-  array_init_(&crr, arr->n + 1, usize, __FILE__, __LINE__);
-
-  char *pe = (char *)arr->ptr;
-  uint off = 0;
-  for (unsigned i = 0; i < nt; i++) {
-    // Copy from arr to brr.
-    brr.n = 0;
-    uint off1 = off + tsize;
-    for (uint j = off; j < arr->n && j < off1; j++)
-      array_cat_(usize, &brr, &pe[j * usize], 1, __FILE__, __LINE__);
-    sarray_transfer_ext_(&brr, usize, &proc[off], sizeof(uint), &cr);
-    array_cat_(usize, &crr, brr.ptr, brr.n, __FILE__, __LINE__);
-    off = (off1 < arr->n ? off1 : arr->n);
-  }
-  array_free(&brr), free(proc);
-
-  array_free(arr), arr->n = 0;
-  array_cat_(usize, arr, crr.ptr, crr.n, __FILE__, __LINE__);
-  array_free(&crr);
-
-  crystal_free(&cr);
-
-  // Locally sort again to make sure that we have both globally and locally
-  // sorted array.
-  sort_local(s);
-}
-
-//-----------------------------------------------------------------------------
-// Parallel Hypercube-Sort
-//
-struct hypercube {
-  struct sort *data;
-  int nprobes;
-  double *probes;
-  ulong *probe_cnt;
-};
-
-static int init_probes(struct hypercube *data, struct comm *c) {
-  struct sort *input = data->data;
-
-  // Allocate space for probes and counts
-  int nprobes = data->nprobes = 3;
-  if (!data->probes)
-    data->probes = tcalloc(double, nprobes);
-  if (!data->probe_cnt)
-    data->probe_cnt = tcalloc(ulong, nprobes);
-
-  double extrema[2];
-  get_extrema((void *)extrema, data->data, 0, c);
-  double range = extrema[1] - extrema[0];
-  double delta = range / (nprobes - 1);
-
-  data->probes[0] = extrema[0];
-  data->probes[1] = extrema[0] + delta;
-  data->probes[2] = extrema[1];
-
-  return 0;
-}
-
-static int update_probe_counts(struct hypercube *data, struct comm *c) {
-  struct sort *input = data->data;
-  uint offset = input->offset[0];
-  gs_dom t = input->t[0];
-
-  uint nprobes = data->nprobes;
-  uint i;
-  for (i = 0; i < nprobes; i++)
-    data->probe_cnt[i] = 0;
-
-  struct array *a = input->a;
-  uint e;
-  for (e = 0; e < a->n; e++) {
-    double val_e = get_scalar(a, e, offset, input->unit_size, t);
-    for (i = 0; i < nprobes; i++)
-      if (val_e < data->probes[i])
-        data->probe_cnt[i]++;
-  }
-
-  slong buf[6];
-  comm_allreduce(c, gs_long, gs_add, data->probe_cnt, nprobes, buf);
-
-  return 0;
-}
-
-static int update_probes(slong nelem, double *probes, ulong *probe_cnt,
-                         uint threshold) {
-  slong expected = nelem / 2;
-  if (llabs(expected - (slong)probe_cnt[1]) < threshold)
-    return 0;
-
-  if (probe_cnt[1] < expected)
-    probes[0] = probes[1];
-  else
-    probes[2] = probes[1];
-
-  probes[1] = probes[0] + (probes[2] - probes[0]) / 2.0;
-
-  return 0;
-}
-
-static int transfer_elem(struct hypercube *data, struct comm *c) {
-  struct sort *input = data->data;
-  uint usize = input->unit_size, offset = input->offset[0];
-  gs_dom t = input->t[0];
-  struct array *a = input->a;
-
-  uint size = a->n, lown = 0, uppern = 0;
-  for (uint e = 0; e < size; e++) {
-    double val = get_scalar(a, e, offset, usize, t);
-    if (val < data->probes[1])
-      lown++;
-    else
-      uppern++;
-  }
-
-  slong out[2][2], in[2] = {lown, uppern}, buf[2][2];
-  comm_scan(out, c, gs_long, gs_add, in, 2, buf);
-  slong lstart = out[0][0], ustart = out[0][1];
-  slong lelem = out[1][0], uelem = out[1][1];
-
-  uint np = c->np, lnp = np / 2;
-  uint *proc = tcalloc(uint, size);
-  set_dest(proc, lnp, lstart, lown, lelem);
-  set_dest(proc + lown, np - lnp, ustart, uppern, uelem);
-
-  for (uint e = lown; e < size; e++)
-    proc[e] += lnp;
-
-  struct crystal cr;
-  crystal_init(&cr, c);
-  sarray_transfer_ext_(a, usize, proc, sizeof(uint), &cr);
-  crystal_free(&cr);
-
-  free(proc);
-
-  return 0;
-}
-
-static int parallel_hypercube_sort(struct hypercube *data, struct comm *c) {
-  struct sort *input = data->data;
-  struct array *a = input->a;
-  gs_dom t = input->t[0];
-  uint offset = input->offset[0];
-
-  sint size = c->np, rank = c->id;
-
-  slong out[2][1], buf[2][1], in = a->n;
-  comm_scan(out, c, gs_long, gs_add, &in, 1, buf);
-  slong start = out[0][0];
-  slong nelem = out[1][0];
-
-  uint threshold = nelem / (10 * size);
-  if (threshold < 2)
-    threshold = 2;
-
-  sort_local(data->data);
-
-  if (size == 1)
-    return 0;
-
-  init_probes(data, c);
-  update_probe_counts(data, c);
-
-  int max_iter = log2((data->probes[2] - data->probes[0]) / 1e-12);
-  int iter = 0;
-  while (llabs(nelem / 2 - (slong)data->probe_cnt[1]) > threshold &&
-         iter++ < max_iter) {
-    update_probes(nelem, data->probes, data->probe_cnt, threshold);
-    update_probe_counts(data, c);
-  }
-
-  transfer_elem(data, c);
-
-  // split the communicator
-  struct comm nc;
-  sint lower = (rank < size / 2) ? 1 : 0;
-#if defined(MPI)
-  MPI_Comm nc_;
-  MPI_Comm_split(c->c, lower, rank, &nc_);
-  comm_init(&nc, nc_);
-  MPI_Comm_free(&nc_);
-#else
-  comm_init(&nc, 1);
-#endif
-
-  // TODO: Keep load balancing after each split
-  parallel_hypercube_sort(data, &nc);
-  comm_free(&nc);
-
-  return 0;
 }
 
 static int load_balance(struct array *a, size_t size, const struct comm *c,
@@ -371,14 +111,14 @@ static int load_balance(struct array *a, size_t size, const struct comm *c,
   slong start = out[0][0], nelem = out[1][0];
 
   uint *proc = tcalloc(uint, a->n);
-  set_dest(proc, a->n, c->np, start, nelem);
+  set_proc_from_idx(proc, a->n, c->np, start, nelem);
   sarray_transfer_ext_(a, size, proc, sizeof(uint), cr);
   free(proc);
 
   return 0;
 }
 
-int parallel_sort_private(struct sort *data, const struct comm *c) {
+void parallel_sort_private(struct sort *data, const struct comm *c) {
   struct comm dup;
   comm_dup(&dup, c);
 
@@ -411,6 +151,4 @@ int parallel_sort_private(struct sort *data, const struct comm *c) {
     sort_local(data);
   }
   comm_free(&dup);
-
-  return 0;
 }
