+++ src/sort.h
@@ -16,10 +16,8 @@ struct sort {
   buffer *buf;
 };
 
-int sort_local(struct sort *s);
-int parallel_sort_private(struct sort *s, const struct comm *c);
+void parallel_sort_private(struct sort *s, const struct comm *c);
 
-// Uniform parallel sort
 #define parallel_sort(T, A, field, type, method, loadbalance, c, bufp)         \
   do {                                                                         \
     struct sort sd;                                                            \
